almost complete

+++ pursuit_backend/pursuit_app/models/Interview.py
@@ -6,7 +6,7 @@ class Interview(models.Model) :
 
     rounds = models.ForeignKey( Round , on_delete = models.CASCADE , )
     student = models.ForeignKey( Student , on_delete = models.CASCADE , related_name='interviews')
-    start_time = models.PositiveIntegerField('start time')
+    start_time = models.PositiveBigIntegerField('start time')
 
     def __str__(self) :
         return  str(self.id)
